+++ check_cols.py
@@ -1,0 +1,4 @@
+import akshare as ak
+df = ak.stock_news_em(symbol="600076")
+print(df.columns)
+print(df.iloc[0])